+++ example/src/App.svelte
@@ -2,8 +2,8 @@
 	import Typewriter from 'Typewriter.svelte'
 </script>
 
-<Typewriter cascade on:done={() => console.log('Hello World!')}>
+<Typewriter on:done={() => console.log('Hello World!')}>
   <p>Lorem & Ipsum</p>
   <p>Dolor & Sit</p>
-  <p>Amet & Consectetur</p>
+  <p>Amet & Consectetur... And a lot of some random text just to test if my code works</p>
 </Typewriter>